backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-multipart>=0.0.5
numpy>=1.21.2
redis>=4.5.1
aioredis>=2.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
requests>=2.26.0
matplotlib>=3.4.3
